Build University Layer

+++ xhdh/xhdh/src/main/java/com/xhdh/xhdh/models/UniversityTag.java
@@ -4,8 +4,6 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.io.Serializable;
-
 @Getter
 @Setter
 @Entity
@@ -14,5 +12,5 @@ public class UniversityTag {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int universityId;
-    private byte tagId;
+    private int tagId;
 }
